data/scrapers/requirements.txt
# Core dependencies
PyYAML>=6.0.0
beautifulsoup4>=4.12.0
aiosqlite>=0.19.0
typer>=0.12.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
structlog>=24.1.0
PyMuPDF>=1.24.0

# Browser engine
playwright>=1.44.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
